+++ modal.js
@@ -3,7 +3,7 @@
  * http://drublic.github.com/css-modal
  *
  * @author Hans Christian Reinl - @drublic
- * @version 1.0.1
+ * @version 1.0.4
  */
 
 (function (global) {
@@ -18,13 +18,16 @@
 	modal.activeElement = undefined;
 
 	// Polyfill addEventListener for IE8 (only very basic)
-	document._addEventListener = document.addEventListener ||
-		function (event, callback) {
-			document.attachEvent('on' + event, callback);
-		};
+	modal._addEventListener = function (element, event, callback) {
+		if (element.addEventListener) {
+			element.addEventListener(event, callback, false);
+		} else {
+			element.attachEvent('on' + event, callback);
+		}
+	};
 
 	// Hide overlay when ESC is pressed
-	document._addEventListener('keyup', function (event) {
+	modal._addEventListener(document, 'keyup', function (event) {
 		var hash = window.location.hash.replace('#', '');
 
 		// If hash is not set
@@ -63,11 +66,12 @@
 
 
 	// When showing overlay, prevent background from scrolling
-	window.onhashchange = function () {
+	modal.mainHandler = function () {
 		var hash = window.location.hash.replace('#', '');
 		var modalElement = document.getElementById(hash);
 		var htmlClasses = document.documentElement.className;
 		var modalChild;
+		var oldModal;
 
 		// If the hash element exists
 		if (modalElement) {
@@ -76,12 +80,18 @@
 			modalChild = modalElement.children[0];
 
 			// When we deal with a modal and body-class `has-overlay` is not set
-			if (modalChild && modalChild.className.match(/modal-inner/) &&
-					!htmlClasses.match(/has-overlay/)) {
+			if (modalChild && modalChild.className.match(/modal-inner/)) {
+				if (!htmlClasses.match(/has-overlay/)) {
 
-				// Set an html class to prevent scrolling
-				document.documentElement.className += ' has-overlay';
+					// Set an html class to prevent scrolling
+					document.documentElement.className += ' has-overlay';
+				}
 
+				// Unmark previous active element
+				if (modal.activeElement) {
+					oldModal = modal.activeElement;
+					oldModal.className = oldModal.className.replace(' is-active', '');
+				}
 				// Mark modal as active
 				modalElement.className += ' is-active';
 				modal.activeElement = modalElement;
@@ -113,6 +123,8 @@
 		}
 	};
 
+	modal._addEventListener(window, 'hashchange', modal.mainHandler);
+	modal._addEventListener(window, 'load', modal.mainHandler);
 
 	/*
 	 * Accessibility