+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 
+* Bugfix: UTF-8 problem in Ruby 1.9.3 (#61)
 * Bugfix: open modal on pageload (#46)
 * Register CSSModal as module for AMD (#57)
 * Prevent body from scrolling when scrolling in modal on iOS and Android (#31)